+++ src/harlequin/components/__init__.py
@@ -1,5 +1,5 @@
 from harlequin.components.code_editor import CodeEditor, EditorCollection
-from harlequin.components.data_catalog import CatalogItem, DataCatalog
+from harlequin.components.data_catalog import DataCatalog
 from harlequin.components.error_modal import ErrorModal
 from harlequin.components.export_screen import ExportScreen, export_callback
 from harlequin.components.help_screen import HelpScreen
@@ -15,6 +15,5 @@ __all__ = [
     "HelpScreen",
     "ResultsViewer",
     "RunQueryBar",
-    "CatalogItem",
     "DataCatalog",
 ]